Use Rubocop's Rubocop::RSpec::ExpectOffense for better specs
We had previously copied in the helper from Rubocop a few years ago. This updates us to the latest file in Rubocop itself, which is much improved. There's no reason to vendor this since we pin to specific rubocop releases and we can either vendor or update specs if things change in the future.

Changes
	- Gives us nice helpers for testing the autocorrects which we really need
	- Renames Violations to be Offenses

Signed-off-by: Tim Smith <tsmith@chef.io>