Add Chef/NodeMethodsInsteadofAttributes cop
This detects changes due our the removal of the node method missing behavior in Chef 13. I'm not calling it that because no one knows what method missing is who just needs to clean up some cookbooks. This catches the bad methods we can be 100% sure of. There's 250 of these on the supermarket. Obviously it's not going to catch all your cookbook attributes, but it gets you a bit closer.

Signed-off-by: Tim Smith <tsmith@chef.io>