Add ChefStyle/SimplifyPlatformMajorVersionCheck
When checking the major version number of a platform you can take the node['platform_version'] attribute and transform it to an integer to strip it down to just the major version number. This simple way of determining the major version number of a platform should be used instead of splitting the platform into multiple fields with '.' as the delimiter.

  # bad
  node['platform_version'].split('.').first
  node['platform_version'].split('.')[0]
  node['platform_version'].split('.').first.to_i
  node['platform_version'].split('.')[0].to_i

  # good

  if node['platform_version'].to_i == 7

Signed-off-by: Tim Smith <tsmith@chef.io>