Add new TargetChefVersion config option
This allows the user to disable any cop that would suggest updating a
cookbook in a way that would make it incompatible with earlier versions
of Chef Infra. This is necessary to allow users to step upgrade from 12
-> 13 -> 14 -> 15 without breaking a cookbook in a way where it will
only work in Chef Infra 15. Most of this is monkeypatching rubocop
itself to add / modify methods the same way Rubocop core already
supports this code with Rails. Lots of patching. Ideally Rubocop can do
this natively without the monkeypatching soon.

Signed-off-by: Tim Smith <tsmith@chef.io>