Rework how we autocorrect allowed actions
Don't remove them if we're not in a custom resource
Consider the custom resource cop to be ChefRedundant since there has never been a need for this code in a custom resource. It's just leftover redundant junk
Add a new cop ChefModernize/AllowedActionsFromInitialize that pulls the allowed_actions out of initialize methods in HWRPs. This gets folks closer to LWRPs which gets them closer to custom resources

Signed-off-by: Tim Smith <tsmith@chef.io>