Handle more multipackage scenarios
Detect when blocks that include more than just the package install. It's pretty common that they also include an apt_update or an include_recipe 'apt' when we're on a debian-ish platform.

I plan to open another PR to refactor the if detection with this same support. baby steps here.

Signed-off-by: Tim Smith <tsmith@chef.io>