Fix MetadataMissingName to properly autocorrect
Right now MetadataMissingName is firing anytime it scans. This means if
you have a non cookbook project that has a file named metadata.rb and
you use the chef vscode extension that file will be modified just by
opening it up in vscode. This keeps happening to me in omnibus. Convert
this to use native correct calls.

Signed-off-by: Tim Smith <tsmith@chef.io>