Add chefstyle library file for Rake backwards compatibility
There's a decent amount of Rake code that uses `require "chefstyle"` to set the
defaults for RuboCop (eg
https://github.com/chef/ohai/blob/6d64237f9987c3bf51805e19884e6e710c3a40f6/Rakefile#L19),
and rather than doing surgery on each repo's Rakefile, it'd be easier to keep
that same behavior for loading the configuration and cops. With this change
a migration involving Rake should only involve a Gemfile update and changing
`require "chefstyle"` to `require "cookstyle/chefstyle"`.

Signed-off-by: David Crosby <dcrosby@fb.com>